+++ pact-js-master/examples/ava/test/get-dog_1.spec.js
@@ -14,8 +14,8 @@ const provider = new Pact({
   log: path.resolve(process.cwd(), "logs", "mockserver-integration.log"),
   dir: path.resolve(process.cwd(), "pacts"),
   spec: 2,
-  consumer: "MyConsumer",
-  provider: "MyProvider",
+  consumer: "Ava Consumer Example",
+  provider: "Ava Provider Example",
   pactfileWriteMode: "merge",
 })
 
@@ -71,13 +71,11 @@ test("Dog API GET /dogs/1", async t => {
       name: "rocky",
     },
   ])
-})
 
-test.afterEach(async t => {
   // verify with Pact, and reset expectations
-  await t.notThrows(provider.verify())
+  await provider.verify()
 })
 
-test.always.after("pact.js mock server graceful shutdown", async () => {
+test.after.always("pact.js mock server graceful shutdown", async () => {
   await provider.finalize()
 })