+++ pact-js-master/src/messageConsumerPact.ts
@@ -11,7 +11,7 @@ import {
   MessageProvider,
   MessageConsumer,
 } from "./dsl/message"
-import logger from "./common/logger"
+import logger, { setLogLevel } from "./common/logger"
 import serviceFactory from "@pact-foundation/pact-node"
 import { MessageConsumerOptions } from "./dsl/options"
 import ConfigurationError from "./errors/configurationError"
@@ -30,6 +30,7 @@ export class MessageConsumerPact {
 
   constructor(private config: MessageConsumerOptions) {
     if (!isEmpty(config.logLevel)) {
+      setLogLevel(config.logLevel)
       serviceFactory.logLevel(config.logLevel)
     }
   }